restructure code and start improving test coverage
