+++ yupana/processor/report_processor.py
@@ -1,0 +1,801 @@
+#
+# Copyright 2019 Red Hat, Inc.
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+#
+"""Upload message report processor."""
+
+import asyncio
+import base64
+import json
+import logging
+import tarfile
+import threading
+from datetime import datetime
+from enum import Enum
+from http import HTTPStatus
+from io import BytesIO
+
+import requests
+from aiokafka import AIOKafkaProducer
+from django.db import transaction
+from kafka.errors import ConnectionError as KafkaConnectionError
+from processor.kafka_msg_handler import (KafkaMsgHandlerError,
+                                         QPCKafkaMsgException,
+                                         QPCReportException,
+                                         format_message)
+
+from api.models import Report, ReportArchive
+from config.settings.base import (INSIGHTS_HOST_INVENTORY_URL,
+                                  INSIGHTS_KAFKA_ADDRESS)
+
+LOG = logging.getLogger(__name__)
+PROCESSING_LOOP = asyncio.new_event_loop()
+VALIDATION_TOPIC = 'platform.upload.validation'
+SUCCESS_CONFIRM_STATUS = 'success'
+FAILURE_CONFIRM_STATUS = 'failure'
+CANONICAL_FACTS = ['insights_client_id', 'bios_uuid', 'ip_addresses', 'mac_addresses',
+                   'vm_uuid', 'etc_machine_id', 'subscription_manager_id']
+
+RETRY_TIME = 1  # this is the time in minutes that we want to wait to retry a report
+RETRIES_ALLOWED = 5  # this is the number of retries that we want to allow before failing a report
+FAILED_VERIFICATION = 'VERIFICATION'
+FAILED_UPLOAD = 'UPLOAD'
+EMPTY_QUEUE_SLEEP = 60
+RETRY = Enum('RETRY', 'clear increment archive')
+
+
+class MessageProcessor():
+    """Class for processing saved upload service messages."""
+
+    def __init__(self):
+        """Create a message processor."""
+        self.report = None
+        self.state = None
+        self.next_state = None
+        self.account_number = None
+        self.upload_message = None
+        self.report_id = None
+        self.report_json = None
+        self.candidate_hosts = None
+        self.failed_hosts = None
+        self.status = None
+        self.prefix = 'PROCESSING REPORT'
+
+    async def run(self):
+        """Run the message processor for a report."""
+        while True:
+            if not self.report:
+                self.assign_report()
+            if self.report:
+                try:
+                    await self.delegate_state()
+                except Exception as error:
+                    LOG.error(format_message(self.prefix,
+                                             'The following error occurred: %s.',
+                                             str(error)))
+            else:
+                await asyncio.sleep(EMPTY_QUEUE_SLEEP)
+
+    @transaction.atomic
+    def assign_report(self):
+        """Assign the message processor reports that are saved in the db.
+
+        First priority is the oldest reports that are in the new state.
+        If no reports meet this condition, we look for the oldest report in any state.
+        We then check to see if an appropriate amount of time has passed before we retry this
+        report.
+        """
+        self.prefix = 'ASSIGNING REPORT'
+        report_found_message = 'Starting report processor. State is "%s"'
+        if self.report is None:
+            try:
+                # look for the oldest report in the db
+                oldest_report = Report.objects.earliest('last_update_time')
+                current_time = datetime.utcnow()
+                minutes_passed = int(
+                    (current_time - oldest_report.last_update_time).total_seconds() / 60)
+                if minutes_passed >= RETRY_TIME:
+                    self.report = oldest_report
+                    self.next_state = Report.STARTED
+                    LOG.info(format_message(
+                        self.prefix, report_found_message % self.report.state,
+                        account_number=self.account_number, report_id=self.report_id))
+                    self.update_report_state()
+            except Report.DoesNotExist:
+                try:
+                    # otherwise, look for the oldest report in the new state
+                    self.report = Report.objects.filter(
+                        state=Report.NEW).earliest('last_update_time')
+                    LOG.info(format_message(self.prefix, report_found_message % self.report.state,
+                                            account_number=self.account_number))
+                    # update the report time and state
+                    self.next_state = Report.STARTED
+                    self.update_report_state()
+                except Report.DoesNotExist:
+                    report_not_found_message = \
+                        'No reports to be processed at this time. '\
+                        'Checking again in % s seconds.' % str(EMPTY_QUEUE_SLEEP)
+                    LOG.info(format_message(self.prefix, report_not_found_message))
+
+    async def delegate_state(self):
+        """Call the correct function based on report state.
+
+        If the function is async, make sure to await it.
+        """
+        self.state = self.report.state
+        self.account_number = self.report.rh_account
+        self.upload_message = json.loads(self.report.upload_srv_kafka_msg)
+        if self.report.candidate_hosts:
+            self.candidate_hosts = json.loads(self.report.candidate_hosts)
+        if self.report.failed_hosts:
+            self.failed_hosts = json.loads(self.report.failed_hosts)
+        if self.report.report_json:
+            self.report_json = json.loads(self.report.report_json)
+        if self.report.report_platform_id:
+            self.report_id = self.report.report_platform_id
+        if self.report.upload_ack_status:
+            self.status = self.report.upload_ack_status
+        async_function_call_states = [Report.VALIDATED]
+        state_functions = {Report.STARTED: self.transition_to_download,
+                           Report.DOWNLOADED: self.transition_to_validate,
+                           Report.VALIDATED: self.transition_to_validation_reported,
+                           Report.VALIDATION_REPORTED: self.transition_to_hosts_uploaded,
+                           Report.HOSTS_UPLOADED: self.archive_report,
+                           Report.FAILED_DOWNLOAD: self.archive_report,
+                           Report.FAILED_VALIDATION: self.archive_report,
+                           Report.FAILED_VALIDATION_REPORTING: self.archive_report,
+                           Report.FAILED_HOSTS_UPLOAD: self.archive_report}
+        # if the function is async, we must await it
+        if state_functions.get(self.state):
+            if self.state in async_function_call_states:
+                await state_functions.get(self.state)()
+            else:
+                state_functions.get(self.state)()
+        else:
+            LOG.error(format_message(
+                self.prefix,
+                'An error occurred that caused this report to fall into state "%s".' % self.state,
+                account_number=self.account_number,
+                report_id=self.report_id))
+
+    def transition_to_download(self):
+        """Attempt to download the report, extract the json & move to downloaded state."""
+        self.prefix = 'ATTEMPTING DOWNLOAD'
+        LOG.info(format_message(
+            self.prefix,
+            'Attempting to download the report and extract the json. '
+            'State is "%s"' % self.report.state,
+            account_number=self.account_number))
+        try:
+            report_tar_gz = self._download_report()
+            self.report_json = self._extract_report_from_tar_gz(report_tar_gz)
+            self.next_state = Report.DOWNLOADED
+            self.update_report_state(report_json=self.report_json)
+        except (QPCReportException, QPCKafkaMsgException):
+            self.determine_retry(Report.FAILED_DOWNLOAD, Report.STARTED)
+
+    def transition_to_validate(self):
+        """Validate that the report contents & move to validated state."""
+        self.prefix = 'ATTEMPTING VALIDATE'
+        LOG.info(format_message(
+            self.prefix, 'Validating the report contents. State is "%s"' % self.report.state,
+            account_number=self.account_number))
+        try:
+            self.candidate_hosts, self.failed_hosts = self._validate_report_details()
+            failed_hosts_list = self.assign_cause_to_failed(FAILED_VERIFICATION)
+            self.report_id = self.report_json.get('report_platform_id')
+            self.status = SUCCESS_CONFIRM_STATUS
+            self.next_state = Report.VALIDATED
+            if not self.candidate_hosts:
+                self.status = FAILURE_CONFIRM_STATUS
+            self.update_report_state(report_id=self.report_id, candidate_hosts=self.candidate_hosts,
+                                     failed_hosts=failed_hosts_list, status=self.status)
+            self.deduplicate_reports()
+        except Exception as error:
+            LOG.error(format_message(self.prefix, 'The following error occurred: %s.', str(error)))
+            self.determine_retry(Report.FAILED_VALIDATION, Report.DOWNLOADED)
+
+    async def transition_to_validation_reported(self):
+        """Upload the validation status & move to validation reported state."""
+        self.prefix = 'ATTEMPTING STATUS UPLOAD'
+        LOG.info(format_message(
+            self.prefix,
+            'Uploading validation status %s for report %s' % (self.status, self.report_id),
+            account_number=self.account_number, report_id=self.report_id))
+        message_hash = self.upload_message['hash']
+        try:
+            await self._send_confirmation(message_hash)
+            self.next_state = Report.VALIDATION_REPORTED
+            self.update_report_state()
+        except Exception as error:
+            LOG.error(format_message(self.prefix, 'The following error occurred: %s.' % str(error),
+                                     account_number=self.account_number, report_id=self.report_id))
+            self.determine_retry(Report.FAILED_VALIDATION_REPORTING, Report.VALIDATED)
+
+    def transition_to_hosts_uploaded(self):
+        """Upload the host candidates to inventory & move to hosts_uploaded state."""
+        self.prefix = 'ATTEMPTING HOST UPLOAD'
+        LOG.info(format_message(
+            self.prefix, 'Uploading hosts to inventory. State is "%s"' % self.report.state,
+            account_number=self.account_number, report_id=self.report_id))
+        try:
+            hosts_to_try = self.generate_candidates()
+            if hosts_to_try:
+                self.candidate_hosts, self.failed_hosts = \
+                    self._upload_to_host_inventory(hosts_to_try)
+                self.next_state = Report.HOSTS_UPLOADED
+                failed_hosts_list = self.assign_cause_to_failed(FAILED_UPLOAD)
+                if self.candidate_hosts:
+                    self.update_report_state(candidate_hosts=self.candidate_hosts,
+                                             failed_hosts=failed_hosts_list)
+                else:
+                    self.determine_retry(Report.FAILED_HOSTS_UPLOAD, Report.VALIDATION_REPORTED)
+            else:
+                # need to not upload, but archive bc no hosts were valid
+                LOG.info(format_message(self.prefix, 'There are no valid hosts to upload',
+                                        account_number=self.account_number,
+                                        report_id=self.report_id))
+                self.archive_report()
+        except Exception as error:
+            LOG.error(format_message(self.prefix, 'The following error occurred: %s.' % str(error),
+                                     account_number=self.account_number, report_id=self.report_id))
+            self.determine_retry(Report.FAILED_HOSTS_UPLOAD, Report.VALIDATION_REPORTED)
+
+    @transaction.atomic
+    def archive_report(self):
+        """Archive the report object."""
+        self.prefix = 'ARCHIVING REPORT'
+        LOG.info(format_message(self.prefix, 'Archiving report.',
+                                account_number=self.account_number, report_id=self.report_id))
+        archived = ReportArchive(
+            rh_account=self.account_number,
+            retry_count=self.report.retry_count,
+            candidate_hosts=self.report.candidate_hosts,
+            failed_hosts=self.report.failed_hosts,
+            state=self.state,
+            state_info=self.report.state_info,
+            last_update_time=self.report.last_update_time,
+            upload_srv_kafka_msg=self.upload_message
+        )
+        if self.report_id:
+            archived.report_platform_id = self.report_id
+        if self.report_json:
+            archived.report_json = self.report_json
+        if self.status:
+            archived.upload_ack_status = self.status
+        archived.save()
+        try:
+            Report.objects.get(id=self.report.id).delete()
+        except Report.DoesNotExist:
+            pass
+        LOG.info(format_message(self.prefix, 'Report successfully archived.',
+                                account_number=self.account_number, report_id=self.report_id))
+        self.reset_variables()
+
+    def update_report_state(self, retry=RETRY.clear, report_json=None, report_id=None,  # noqa: C901 (too-complex)
+                            candidate_hosts=None, failed_hosts=None, status=None):
+        """
+        Update the message processor state and save.
+
+        :param retry: <bool> boolean on whether or not we should increase the retry
+        :param report_json: <dict> dictionary containing the report json
+        :param report_id: <str> string containing report_platform_id
+        :param candidate_hosts: <dict> dictionary containing hosts that were
+            successfully verified and uploaded
+        :param failed_hosts: <dict> dictionary containing hosts that failed
+            verification or upload
+        """
+        self.prefix = 'UPDATING REPORT STATE'
+        try:
+            self.report.last_update_time = datetime.utcnow()
+            self.report.state = self.next_state
+            if retry == RETRY.clear:
+                # reset the count to 0 (default behavior)
+                self.report.retry_count = 0
+            elif retry == RETRY.increment:
+                self.report.retry_count += 1
+            # the other choice for retry is RETRY.archive in which case we don't
+            # want to do anything to the retry count bc we want to preserve as is
+            if report_json:
+                self.report.report_json = json.dumps(report_json)
+            if report_id:
+                self.report.report_platform_id = report_id
+            if candidate_hosts:
+                # for success hosts, these can change based on the function
+                # ie. hosts may pass verification but not upload so we
+                # completely override the previous value to get only the
+                # successful hosts at that point. Also, we must remove the
+                # hosts that succeeded from the failed hosts in case they were
+                # put there on a retry.
+                self.remove_success_from_failure(candidate_hosts)
+                self.report.candidate_hosts = json.dumps(candidate_hosts)
+            if failed_hosts:
+                # for failed hosts this list can keep growing, so we add the
+                # newly failed hosts to the previous value
+                failed = json.loads(self.report.failed_hosts)
+                for host in failed_hosts:
+                    failed.append(host)
+                self.report.failed_hosts = json.dumps(failed)
+            if status:
+                self.report.upload_ack_status = status
+            state_info = json.loads(self.report.state_info)
+            state_info.append(self.next_state)
+            self.report.state_info = json.dumps(state_info)
+            self.report.save()
+            LOG.info(format_message(self.prefix, 'Report successfully updated.',
+                                    account_number=self.account_number, report_id=self.report_id))
+        except Exception as error:
+            LOG.error(format_message(
+                self.prefix,
+                'Could not update report record due to the following error %s.' % str(error),
+                account_number=self.account_number, report_id=self.report_id))
+
+    def determine_retry(self, fail_state, current_state, failed_hosts=None):
+        """Determine if yupana should archive a report based on retry count.
+
+        :param fail_state: <str> the final state if we have reached max retries
+        :param current_state: <str> the current state we are in that we want to try again
+        :param failed_hosts: <list> for uploading to host inventory, if we fail due to
+            no hosts being successfully uploaded, we want to rerecord the hosts that failed
+            for retrying.
+        """
+        if self.report.retry_count + 1 > RETRIES_ALLOWED:
+            LOG.error(format_message(
+                self.prefix, 'This report has reached the maximum number of retries.',
+                account_number=self.account_number, report_id=self.report_id))
+            self.next_state = fail_state
+            self.update_report_state(retry=RETRY.archive)
+        else:
+            self.next_state = current_state
+            LOG.error(format_message(
+                self.prefix, 'An error occurred. Saving report to retry later.',
+                account_number=self.account_number, report_id=self.report_id))
+            if failed_hosts:
+                self.update_report_state(
+                    retry=RETRY.increment, candidate_hosts=self.candidate_hosts,
+                    failed_hosts=failed_hosts)
+            else:
+                self.update_report_state(retry=RETRY.increment)
+            self.reset_variables()
+
+    def assign_cause_to_failed(self, cause):
+        """Assign the reason for failure to the failed_hosts.
+
+        :param cause: <str> should be 'VERIFICATION' or 'UPLOAD'
+        """
+        failed_hosts_list = []
+        for host_id, host in self.failed_hosts.items():
+            failed_hosts_list.append({host_id: host, 'cause': cause})
+        return failed_hosts_list
+
+    def generate_candidates(self):
+        """Generate hosts that need to be uploaded to host inventory.
+
+        If a retry has not occurred then we return the candidate_hosts
+        but if a retry has occurred and failed at uploading, we want to retry
+        the hosts that failed upload while excluding the ones that succeeded.
+        """
+        failed_hosts_list = json.loads(self.report.failed_hosts)
+        success_hosts = json.loads(self.report.candidate_hosts)
+        retry_hosts = {}
+        for host in failed_hosts_list:
+            if host.get('cause', '') == FAILED_UPLOAD:
+                host.pop('cause')
+                for key in host.keys():
+                    retry_hosts[key] = host[key]
+        if retry_hosts:
+            return retry_hosts
+        return success_hosts
+
+    def remove_success_from_failure(self, success_hosts):
+        """Remove the hosts that succeeded from the failed hosts if they are there.
+
+        :param success_hosts: <dict> dictionary of successful hosts at this point.
+        """
+        failed_hosts = json.loads(self.report.failed_hosts)
+        for host_id, host in success_hosts.items():
+            for failed_host in failed_hosts:
+                if failed_host.get(host_id):
+                    failed_hosts.remove(failed_host)
+        self.report.failed_hosts = json.dumps(failed_hosts)
+        self.report.save()
+
+    def deduplicate_reports(self):
+        """If a report with the same id already exists, archive the new report."""
+        try:
+            existing_reports = Report.objects.filter(
+                report_platform_id=self.report.report_platform_id)
+            if existing_reports.count() > 1:
+                LOG.error(format_message(
+                    self.prefix,
+                    'a report with the report_platform_id %s already exists.' %
+                    self.report.report_platfrom_id,
+                    account_number=self.account_number, report_id=self.report_id))
+                self.archive_report()
+        except Report.DoesNotExist:
+            pass
+
+    def reset_variables(self):
+        """Reset the class variables to original values."""
+        self.report = None
+        self.state = None
+        self.account_number = None
+        self.upload_message = None
+        self.report_id = None
+        self.report_json = None
+        self.candidate_hosts = None
+        self.failed_hosts = None
+        self.status = None
+        self.prefix = 'PROCESSING REPORT'
+
+    def _download_report(self):
+        """
+        Download report.
+
+        :returns content: The tar.gz binary content or None if there are errors.
+        """
+        self.prefix = 'REPORT DOWNLOAD'
+        try:
+            report_url = self.upload_message.get('url', None)
+            if not report_url:
+                raise QPCReportException(
+                    format_message(
+                        self.prefix,
+                        'kafka message missing report url.  Message: %s' % self.upload_message,
+                        account_number=self.account_number))
+
+            LOG.info(format_message(
+                self.prefix,
+                'downloading %s' % report_url, account_number=self.account_number))
+            download_response = requests.get(report_url)
+            if download_response.status_code != HTTPStatus.OK:
+                raise QPCKafkaMsgException(
+                    format_message(self.prefix,
+                                   'HTTP status code %s returned for URL %s.  Message: %s' % (
+                                       download_response.status_code,
+                                       report_url,
+                                       self.upload_message),
+                                   account_number=self.account_number))
+
+            LOG.info(format_message(
+                self.prefix,
+                'successfully downloaded %s' % report_url,
+                account_number=self.account_number
+            ))
+            return download_response.content
+        except requests.exceptions.HTTPError as err:
+            raise QPCReportException(
+                format_message(self.prefix,
+                               'Unexpected http error for URL %s. Error: %s' % (
+                                   report_url,
+                                   err),
+                               account_number=self.account_number))
+
+    def _extract_report_from_tar_gz(self, report_tar_gz):
+        """Extract Insights report from tar.gz file.
+
+        :param report_tar_gz: A hexstring or BytesIO tarball
+            saved in memory with gzip compression.
+        :returns: Insights report as dict
+        """
+        prefix = 'EXTRACT REPORT FROM TAR'
+        try:
+            tar = tarfile.open(fileobj=BytesIO(report_tar_gz), mode='r:gz')
+            files_check = tar.getmembers()
+            json_files = []
+            for file in files_check:
+                if '.json' in file.name:
+                    json_files.append(file)
+            if len(json_files) > 1:
+                raise QPCReportException(
+                    format_message(prefix,
+                                   'tar.gz contains multiple files.',
+                                   account_number=self.account_number))
+            if json_files:
+                file = json_files[0]
+                tarfile_obj = tar.extractfile(file)
+                report_json_str = tarfile_obj.read().decode('utf-8')
+                try:
+                    insights_report = json.loads(report_json_str)
+                    LOG.info(
+                        format_message(
+                            prefix, 'successful',
+                            account_number=self.account_number,
+                            report_id=insights_report.get('report_platform_id')))
+                    return insights_report
+                except ValueError as error:
+                    raise QPCReportException(
+                        format_message(prefix,
+                                       'Report not JSON. Error: %s' % str(error),
+                                       account_number=self.account_number))
+            raise QPCReportException(
+                format_message(prefix,
+                               'Tar contains no JSON files.',
+                               account_number=self.account_number))
+        except QPCReportException as qpc_err:
+            raise qpc_err
+        except Exception as err:
+            raise QPCReportException(
+                format_message(prefix,
+                               'Unexpected error reading tar.gz: %s' % str(err),
+                               account_number=self.account_number))
+
+    def _validate_report_details(self):
+        """
+        Verify that the report contents are a valid Insights report.
+
+        :returns: tuple contain list of valid and invalid hosts
+        """
+        prefix = 'VALIDATE REPORT STRUCTURE'
+        required_keys = ['report_platform_id',
+                         'report_id',
+                         'report_version',
+                         'hosts']
+        report_id = self.report_json.get('report_platform_id')
+
+        if self.report_json.get('report_type') != 'insights':
+            raise QPCReportException(
+                format_message(
+                    prefix,
+                    'Attribute report_type missing or not equal to insights',
+                    account_number=self.account_number,
+                    report_id=report_id))
+
+        missing_keys = []
+        for key in required_keys:
+            required_key = self.report_json.get(key)
+            if not required_key:
+                missing_keys.append(key)
+
+        if missing_keys:
+            missing_keys_str = ', '.join(missing_keys)
+            raise QPCReportException(
+                format_message(
+                    prefix,
+                    'Report is missing required fields: %s.' % missing_keys_str,
+                    account_number=self.account_number,
+                    report_id=report_id))
+
+        # validate hosts is a dictionary
+        invalid_hosts_message = 'Hosts must be a dictionary that is not empty. ' \
+                                'All keys must be strings and all values must be dictionaries.'
+        hosts = self.report_json.get('hosts')
+        if not hosts or not isinstance(hosts, dict):
+            raise QPCReportException(
+                format_message(
+                    prefix,
+                    invalid_hosts_message,
+                    account_number=self.account_number,
+                    report_id=report_id))
+
+        invalid_host_dict_format = False
+        for host_id, host in hosts.items():
+            if not isinstance(host_id, str) or not isinstance(host, dict):
+                invalid_host_dict_format = True
+                break
+
+        if invalid_host_dict_format:
+            raise QPCReportException(
+                format_message(
+                    prefix,
+                    invalid_hosts_message,
+                    account_number=self.account_number,
+                    report_id=report_id))
+
+        valid_hosts, invalid_hosts = self._validate_report_hosts()
+        number_valid = len(valid_hosts)
+        total = number_valid + len(invalid_hosts)
+        LOG.info(format_message(
+            prefix,
+            '%s/%s hosts are valid.' % (
+                number_valid, total),
+            account_number=self.account_number,
+            report_id=report_id
+        ))
+        if not valid_hosts:
+            raise QPCReportException(
+                format_message(
+                    prefix,
+                    'report does not contain any valid hosts.',
+                    account_number=self.account_number,
+                    report_id=report_id))
+        else:
+            return valid_hosts, invalid_hosts
+
+    def _validate_report_hosts(self):
+        """Verify that report hosts contain canonical facts.
+
+        :returns: tuple containing valid & invalid hosts
+        """
+        hosts = self.report_json['hosts']
+        report_id = self.report_json['report_platform_id']
+
+        prefix = 'VALIDATE HOSTS'
+        valid_hosts = {}
+        invalid_hosts = {}
+        for host_id, host in hosts.items():
+            found_facts = False
+            for fact in CANONICAL_FACTS:
+                if host.get(fact):
+                    found_facts = True
+                    break
+            if found_facts:
+                valid_hosts[host_id] = host
+            else:
+                host.pop('metadata', None)
+                invalid_hosts[host_id] = host
+        if invalid_hosts:
+            LOG.warning(
+                format_message(
+                    prefix,
+                    'Removed %d hosts with 0 canonical facts: %s' % (
+                        len(invalid_hosts), invalid_hosts),
+                    account_number=self.account_number,
+                    report_id=report_id))
+
+        return valid_hosts, invalid_hosts
+
+    async def _send_confirmation(self, file_hash):  # pragma: no cover
+        """
+        Send kafka validation message to Insights Upload service.
+
+        When a new file lands for topic 'qpc' we must validate it
+        so that it will be made permanently available to other
+        apps listening on the 'available' topic.
+        :param: file_hash (String): Hash for file being confirmed.
+        :returns None
+        """
+        prefix = 'REPORT VALIDATION STATE ON KAFKA'
+        producer = AIOKafkaProducer(
+            loop=PROCESSING_LOOP, bootstrap_servers=INSIGHTS_KAFKA_ADDRESS
+        )
+        try:
+            await producer.start()
+        except (KafkaConnectionError, TimeoutError):
+            await producer.stop()
+            raise KafkaMsgHandlerError(
+                format_message(
+                    prefix,
+                    'Unable to connect to kafka server.  Closing producer.',
+                    account_number=self.account_number,
+                    report_id=self.report_id))
+        try:
+            validation = {
+                'hash': file_hash,
+                'validation': self.status
+            }
+            msg = bytes(json.dumps(validation), 'utf-8')
+            await producer.send_and_wait(VALIDATION_TOPIC, msg)
+            LOG.info(
+                format_message(
+                    prefix,
+                    'Send %s validation status to file upload on kafka' % self.status,
+                    account_number=self.account_number,
+                    report_id=self.report_id))
+        finally:
+            await producer.stop()
+
+    def _upload_to_host_inventory(self, hosts):
+        """
+        Verify that the report contents are a valid Insights report.
+
+        :param hosts: a list of dictionaries that have been validated.
+        :returns None
+        """
+        prefix = 'UPLOAD TO HOST INVENTORY'
+        identity_string = '{"identity": {"account_number": "%s"}}' % str(self.account_number)
+        bytes_string = identity_string.encode()
+        x_rh_identity_value = base64.b64encode(bytes_string).decode()
+        identity_header = {'x-rh-identity': x_rh_identity_value,
+                           'Content-Type': 'application/json'}
+        failed_hosts = []
+        successful_hosts = {}  # storing the hosts that successfully upload here
+        failed_upload_hosts = {}  # storing the hosts that failed to upload here
+        for host_id, host in hosts.items():
+            body = {
+                'account': self.account_number,
+                'bios_uuid': host.get('bios_uuid'),
+                'display_name': host.get('name'),
+                'ip_addresses': host.get('ip_addresses'),
+                'mac_addresses': host.get('mac_addresses'),
+                'insights_id': host.get('insights_client_id'),
+                'rhel_machine_id': host.get('etc_machine_id'),
+                'subscription_manage_id': host.get('subscription_manager_id'),
+                'fqdn': host.get('name'),
+                'facts': [{'namespace': 'qpc',
+                           'facts': host}]
+            }
+            try:
+                response = requests.post(INSIGHTS_HOST_INVENTORY_URL,
+                                         data=json.dumps(body),
+                                         headers=identity_header)
+
+                if response.status_code not in [HTTPStatus.OK, HTTPStatus.CREATED]:
+                    try:
+                        json_body = response.json()
+                    except ValueError:
+                        json_body = 'No JSON response'
+
+                    failed_hosts.append(
+                        {
+                            'status_code': response.status_code,
+                            'error': json_body,
+                            'display_name': body.get('display_name'),
+                            'system_platform_id': host_id,
+                            'host': host})
+                    failed_upload_hosts[host_id] = host
+                else:
+                    successful_hosts[host_id] = host
+
+            except requests.exceptions.RequestException as err:
+                failed_hosts.append(
+                    {
+                        'status_code': 'None',
+                        'error': str(err),
+                        'display_name': body.get('display_name'),
+                        'system_platform_id': host_id,
+                        'host': host})
+                failed_upload_hosts[host_id] = host
+
+        successful = len(hosts) - len(failed_hosts)
+        upload_msg = format_message(
+            prefix, '%s/%s hosts uploaded to host inventory' %
+            (successful, len(hosts)),
+            account_number=self.account_number,
+            report_id=self.report_id
+        )
+        if successful != len(hosts):
+            LOG.warning(upload_msg)
+        else:
+            LOG.info(upload_msg)
+        if failed_hosts:
+            for failed_info in failed_hosts:
+                LOG.error(format_message(
+                    prefix,
+                    'Host inventory returned %s for %s.  Error: %s.  '
+                    'system_platform_id: %s. host: %s' % (
+                        failed_info.get('status_code'),
+                        failed_info.get('display_name'),
+                        failed_info.get('error'),
+                        failed_info.get('system_platform_id'),
+                        failed_info.get('host')),
+                    account_number=self.account_number,
+                    report_id=self.report_id
+                ))
+        return successful_hosts, failed_upload_hosts
+
+
+def asyncio_message_processor_thread(loop):  # pragma: no cover
+    """
+    Worker thread function to run the asyncio event loop.
+
+    :param None
+    :returns None
+    """
+    processor = MessageProcessor()
+    loop.run_until_complete(processor.run())
+
+
+def initialize_message_processor():  # pragma: no cover
+    """
+    Create asyncio tasks and daemon thread to run event loop.
+
+    :param None
+    :returns None
+    """
+    event_loop_thread = threading.Thread(target=asyncio_message_processor_thread,
+                                         args=(PROCESSING_LOOP,))
+    event_loop_thread.daemon = True
+    event_loop_thread.start()
